backend/requirements.txt
flask
flask_sqlalchemy
flask_login
flask_cors
psycopg2-binary